Rename ListViewPage to PooledListView
