+++ Assets/UI/UIPrefab.prefab
@@ -140,7 +140,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1993132893885342}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 30853df40308c4e47a3200e79d738823, type: 3}
+  m_Script: {fileID: 11500000, guid: 36c51c6ac611142f6afdb9c02e367f0a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   ScrollRect: {fileID: 114623274381187898}
